+++ api/core-framework.lua
@@ -1,5 +1,5 @@
 MTH = MTH or {
-	version = "1.4.0",
+	version = "1.4.1",
 	name = "MetaHunt",
 	modules = {},
 	config = {},