+++ modules/chronometer/module.lua
@@ -1,7 +1,7 @@
 local MTH_ChronometerModule = {
 	name = "chronometer",
 	enabled = true,
-	version = "1.4.0",
+	version = "1.4.1",
 	events = {},
 	initialized = false,
 }