+++ modules/icu/module.lua
@@ -6,7 +6,7 @@
 local MTH_ICU = {
 	name = "icu",
 	enabled = false,
-	version = "1.4.0",
+	version = "1.4.1",
 	events = {
 		"VARIABLES_LOADED",
 	},